requirements.txt
pandas
chromadb
sentence-transformers
google-generativeai
mlflow
dagshub
